+++ src/components/indextemplates/IndexTemplateRow.vue
@@ -4,16 +4,14 @@
       <q-icon :name="expand ? 'expand_less' : 'expand_more'" />
     </td>
     <td>{{ row.name }}</td>
-    <td>{{ row.index_patterns }}</td>
-    <td>{{ row.order }}</td>
-    <td>{{ row.version }}</td>
+    <td>{{ row.index_patterns || row.template || row.index_template?.index_patterns }}</td>
   </tr>
 
   <tr v-if="expand">
     <td colspan="100%">
       <div class="q-pa-md">
         <resizable-container>
-          <code-viewer :value="JSON.stringify(row)" />
+          <code-viewer :value="JSON.stringify(cleanedRow)" />
         </resizable-container>
       </div>
     </td>
@@ -21,12 +19,17 @@
 </template>
 
 <script setup lang="ts">
-  import { ref, defineAsyncComponent } from 'vue'
+  import { ref, defineAsyncComponent, computed } from 'vue'
   import ResizableContainer from '../shared/ResizableContainer.vue'
-  import { EsIndexTemplate } from '../../composables/components/indextemplates/IndexTemplatesTable.ts'
+  import { GenericIndexTemplate } from '../../composables/components/indextemplates/IndexTemplates.ts'
 
   const CodeViewer = defineAsyncComponent(() => import('../shared/CodeViewer.vue'))
 
-  defineProps<{ row: EsIndexTemplate }>()
+  const props = defineProps<{ row: GenericIndexTemplate }>()
   const expand = ref(false)
+  const cleanedRow = computed(() => {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const { indexPatterns, ...rest } = props.row
+    return rest
+  })
 </script>